requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
apipkg==3.0.1
apparmor==3.1.3
apsw==3.40.1.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
beautifulsoup4==4.10.0
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cmdln==2.0.0
commonmark==0.9.1
configobj==5.0.6
contourpy==1.0.7
createrepo-c==0.20.1
cryptography==39.0.2
css-parser==1.0.8
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
Django==4.1.7
dnspython==2.3.0
dukpy==0.3
entrypoints==0.4
fastapi==0.94.1
faust-cchardet==2.1.18
feedparser==6.0.10
ffmpy==0.3.0
filelock==3.10.0
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2023.3.0
Genshi==0.7.7
gpg==1.18.0
gradio==3.21.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.2
hyperframe==6.0.1
idna==3.4
ifaddr==0.2.0
inflate64==0.3.1
iniconfig==2.0.0
isodate==0.6.1
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
LibAppArmor==3.1.3
libcomps==0.1.19
linkify-it-py==2.0.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mechanize==0.4.8
msgpack==1.0.4
multidict==6.0.4
multivolumefile==0.2.3
mysqlclient==2.1.1
netifaces==0.11.0
nftables==0.1
notify2==0.3.1
numpy==1.24.2
odfpy==1.4.2
olefile==0.46
openai==0.27.2
orjson==3.8.7
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
py==1.11.0
py7zr==0.20.2
pybcj==1.0.1
pycairo==1.23.0
pychm==0.8.6
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.16.0
pycryptodomex==3.16.0
pycups==2.0.1
pycurl==7.45.2
pydantic==1.10.6
pydub==0.25.1
Pygments==2.14.0
PyGObject==3.42.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyppmd==1.0.0
PyQt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-linux-procfs==0.7.0
python-multipart==0.0.6
pytz==2022.7.1
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0
pyzstd==0.15.3
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.6.0
rpm==4.18.0
sgmllib3k==1.0.0
six==1.16.0
slip==0.6.5
slip.dbus==0.6.5
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.26.1
streamlink==5.3.1
texttable==1.6.7
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.21.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
yarl==1.8.2
zeroconf==0.39.1
